Vendor Management System/vendor_management/requirements.txt
Django
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0